Moved KeyedTimetable to another file
